requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
python-dotenv

# テスト用
pytest
pytest-asyncio
httpx
